refactor: move disable-conflict configs to eslint-plugin-react-x
- Add disable-conflict-eslint-plugin-react and
  disable-conflict-eslint-plugin-react-hooks configs to eslint-plugin-react-x
- Update eslint-plugin to re-export from eslint-plugin-react-x
- Remove no-deprecated preset and all references
