+++ apps/website/content/docs/presets.mdx
@@ -38,11 +38,6 @@ The following presets are available in `@eslint-react/eslint-plugin`:
 - `strict-type-checked`\
   Same as the `strict-typescript` preset but enables additional rules that require type information.
 
-### Category Based
-
-- `no-deprecated`\
-  Enable all rules that disallow deprecated React APIs with "error" severity.
-
 ## Other
 
 - `disable-rsc`\
