+++ packages/plugins/eslint-plugin-react-x/src/index.ts
@@ -1,5 +1,7 @@
 import type { ESLint, Linter } from "eslint";
 
+import * as disableConflictEslintPluginReactConfig from "./configs/disable-conflict-eslint-plugin-react";
+import * as disableConflictEslintPluginReactHooksConfig from "./configs/disable-conflict-eslint-plugin-react-hooks";
 import * as disableExperimentalConfig from "./configs/disable-experimental";
 import * as disableTypeCheckedConfig from "./configs/disable-type-checked";
 import * as recommendedConfig from "./configs/recommended";
@@ -14,6 +16,8 @@ import { plugin } from "./plugin";
 type ConfigName =
   | "disable-experimental"
   | "disable-type-checked"
+  | "disable-conflict-eslint-plugin-react"
+  | "disable-conflict-eslint-plugin-react-hooks"
   | "recommended"
   | "recommended-type-checked"
   | "recommended-typescript"
@@ -32,6 +36,14 @@ const finalPlugin: ESLint.Plugin & { configs: Record<ConfigName, Linter.Config> 
      * Disable rules that can be enforced by TypeScript
      */
     ["disable-type-checked"]: disableTypeCheckedConfig,
+    /**
+     * Disable rules in `eslint-plugin-react` that conflict with rules in this plugin
+     */
+    ["disable-conflict-eslint-plugin-react"]: disableConflictEslintPluginReactConfig,
+    /**
+     * Disable rules in `eslint-plugin-react-hooks` that conflict with rules in this plugin
+     */
+    ["disable-conflict-eslint-plugin-react-hooks"]: disableConflictEslintPluginReactHooksConfig,
     /**
      * Enforce rules that are recommended by ESLint React for general purpose React + React DOM projects
      */
